week_13/ch_2_book_recommendation_app/requirements.txt
streamlit==1.24.0
openai==0.27.8
pinecone-client==2.2.2
google-cloud-translate==3.11.2
google-auth==2.17.3
